Fix bootstrap select that it find the right start and endpoint
The was a bug if you searched first and then select with shift multiple
elements. It was selecting wrong elements.

The new version from upstream doesnt work right and brings new problems
with it. So I decided only to fix the issue with selecting.

There is also an open PR upstream for this issue
https://github.com/davidstutz/bootstrap-multiselect/pull/915